+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 1.0.0
+
+Initial public release. Should be stable enough for others to use.
+
+Added features:
+- Numpad keys to manually trigger actions for testing & demo usage. (for example so the streamer can know what each action does or show to viewers)
+- Information popup when joining a ship, with the option to temporarily disable TiltiSlip for that ship.
+
 ## 0.1.0
 
 Initial coding for stream.